chore(content-switcher): update content-switcher to sb v7
